Improve Font Lock: allow white space and avoid multiple colon
There is no problem for `pass` for handling a field with a white space
on it.  This modification ensures the Font Lock of `pass-view-mode`
indicates that.  As extra modifications, we avoid highlighting after
the first colon.  And for correct highlighting, the colon that
separates the field from its value must be followed by white space.
It also avoids highlighting part of passphrases that contain colon on
it.