Fix byte compile errors
Fix the byte compilation errors:

In toplevel form:
pass.el:51:12: Warning: custom-declare-variable ‘pass-username-fallback-on-filename’ docstring wider than 80 characters

In pass-display-header:
pass.el:354:23: Warning: ‘point-at-bol’ is an obsolete function (as of 29.1); use ‘line-beginning-position’ or ‘pos-bol’ instead.

In pass-display-entry:
pass.el:434:27: Warning: ‘point-at-bol’ is an obsolete function (as of 29.1); use ‘line-beginning-position’ or ‘pos-bol’ instead.

In pass-display-directory:
pass.el:447:27: Warning: ‘point-at-bol’ is an obsolete function (as of 29.1); use ‘line-beginning-position’ or ‘pos-bol’ instead.

In pass-entry-at-point:
pass.el:462:23: Warning: ‘point-at-eol’ is an obsolete function (as of 29.1); use ‘line-end-position’ or ‘pos-eol’ instead.

In pass-otp-from-screenshot:
pass.el:519:2: Warning: docstring wider than 80 characters

In pass-view--otp-counter:
pass.el:661:2: Warning: docstring wider than 80 characters

In end of data:
pass.el:669:28: Warning: the function ‘password-store-otp-token’ is not known to be defined.
pass.el:627:18: Warning: the function ‘password-store-otp-qrcode’ is not known to be defined.